JMAQS #106 - Selenium - CleanUp - 0h 14m -  Moved and Refactored RemoteBrowserType.
Module [JMAQS]:
  * RemoteBrowserType.java
  * SeleniumConfig.java
  * WebDriverFactory.java
  * SeleniumConfigTest.java
  * WebDriverFactoryTest.java
